+++ CHANGELOG.md
@@ -24,12 +24,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement `Sequence` in `retro_memory_map`, `retro_controller_info`,
   `retro_subsystem_rom_info`, and `retro_subsystem_info`.
 
+## Changed
+
+- Removed `boundary=EJECT` from `Key` and `AvEnableFlags`
+  for consistency with other `IntFlag`s.
+
 ### Fixed
 
 - Fixed an error that would be raised when printing a null `c_void_ptr`.
 - Fixed all `Structure` subclasses with pointers
   not converting between null pointers and `None`.
-
 
 ## [0.7.2] - 2026-05-20
 
@@ -59,7 +63,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   and the driver wasn't even implemented.
   It will be readded later.
 - Remove two placeholder guides. Their only content was "TODO".
-
 
 ## [0.7.1] - 2026-04-12